Update to export api auth helpers
